Move env shims to its own module
