+++ src/eval.rs
@@ -39,7 +39,6 @@ pub fn create_ecx<'mir, 'tcx: 'mir>(
         Evaluator::new(config.communicate),
         MemoryExtra::new(StdRng::seed_from_u64(config.seed.unwrap_or(0)), config.validate),
     );
-
     // Complete initialization.
     EnvVars::init(&mut ecx, config.communicate);
